Add bank business date to ForeignExchange model and service
This commit adds the 'bankBusinessDate' property to the ForeignExchange model in order to track the date the exchange rate is fetched from the bank. It also includes the necessary changes in the ForeignExchangeService to create a new ForeignExchange document with the 'bankBusinessDate' value obtained from the BCV_ForeignExchange utility.

Fixes #<issue-number>